Restructure backend folder
